Worked through a bunch of warnings. Some of them were actually warning issues so glad I got through a bunch of them

+++ src/include/scomplexlist.h
@@ -31,7 +31,7 @@ public:
 	void regenerateMoves(void);
 	double getTotalFlux(void);
 	double getJoinFlux(void);
-	uint16_t getMoveCount(void);
+	int getMoveCount(void);
 
 	BaseCount getExposedBases();
 	OpenInfo getOpenInfo();
